int calls/sec

+++ test/combine_stress_results.py
@@ -66,7 +66,7 @@ def format_times(elapsed_float):
         calls_per_second = int((1000 / np.median(t_calls[key])) * n_scripts)
         total_calls_per_second += calls_per_second
 
-        api_calls_per_second = len(t_calls[key]) / elapsed_float
+        api_calls_per_second = int(len(t_calls[key]) / elapsed_float)
         print(f"{prefix} \t {api_calls_per_second} (theoretical max: {calls_per_second})".expandtabs(20))
 
     total_calls = 0
